+++ barelymusician/dsp/dsp_utils_test.cpp
@@ -1,6 +1,7 @@
 #include "barelymusician/dsp/dsp_utils.h"
 
 #include "barelymusician/base/constants.h"
+#include "barelymusician/base/types.h"
 #include "gtest/gtest.h"
 
 namespace barelyapi {
@@ -66,7 +67,7 @@ TEST(DspUtilsTest, GetFilterCoefficient) {
 TEST(DspUtilsTest, SamplesSecondsConversion) {
   const int kSampleRate = 8000;
   const int kNumValues = 4;
-  const int kSamples[kNumValues] = {0, 800, 4000, 32000};
+  const int64 kSamples[kNumValues] = {0, 800, 4000, 32000};
   const double kSeconds[kNumValues] = {0.0, 0.1, 0.5, 4.0};
 
   for (int i = 0; i < kNumValues; ++i) {
